% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_simulations.R
\name{aggregate_simulations}
\alias{aggregate_simulations}
\title{Collapse separate simulation files into a single result}
\usage{
aggregate_simulations(files = NULL, dirs = NULL,
  results_dirname = "SimDesign_aggregate_results")
}
\arguments{
\item{files}{a \code{character} vector containing the names of the simulation files. If \code{NULL},
all files in the working directory ending in \code{.rds} will be used}

\item{dirs}{a \code{character} vector containing the names of the \code{save_results} directories to be
aggregated. A new folder will be created and placed in the \code{results_dirname} output folder}

\item{results_dirname}{the new directory to place the aggregated results files}
}
\value{
if \code{files} is used the function returns a \code{data.frame} with the (weighted) average
  of the simulation results. Otherwise, if \code{dirs} is used, the function returns NULL
}
\description{
This function grabs all \code{.rds} files in the working directory and aggregates them into a single
\code{data.frame} object.
}
\examples{
\dontrun{

setwd('my_working_directory')

## run simulations to save the .rds files (or move them to the working directory)
# runSimulation(..., filename='file1')
# runSimulation(..., filename='file2')

final <- aggregate_simulations()
head(final)

saveRDS(final, 'my_final_simulation.rds')

}
}
\seealso{
\code{\link{runSimulation}}
}

