% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDesign.R
\docType{data}
\name{BF_sim}
\alias{BF_sim}
\title{Example simulation from Brown and Forsythe (1974)}
\description{
Example results from the Brown and Forsythe (1974) article on robust estimators for
variance ratio tests. Statistical tests are organized by columns and the unique design conditoins
are organized by rows. See \code{\link{BF_sim_alternative}} for an alternative form of the same
simulation. Code for this simulation is available of the wiki
(\url{https://github.com/philchalmers/SimDesign/wiki}).
}
\examples{

\dontrun{
data(BF_sim)
head(BF_sim)
}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\references{
Brown, M. B. and Forsythe, A. B. (1974). Robust tests for the equality of variances.
  \emph{Journal of the American Statistical Association, 69}(346), 364--367.
}
\keyword{data}

