% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clusterSetRNGSubStream}
\alias{clusterSetRNGSubStream}
\title{Set RNG sub-stream for  Pierre L'Ecuyer's RngStreams}
\usage{
clusterSetRNGSubStream(cl, seed)
}
\arguments{
\item{cl}{A cluster from the \code{parallel} package, or
(if \code{NULL}) the registered cluster}

\item{seed}{An integer vector of length 7 as given by \code{.Random.seed} when
the L'Ecuyer-CMR RNG is in use. See\code{\link{RNG}} for the valid values}
}
\value{
invisible NULL
}
\description{
Sets the sub-stream RNG state within for Pierre L'Ecuyer's (1999)
algorithm. Should be used within distributed array jobs
after suitable L'Ecuyer's (1999) have been distributed to each array, and
each array is further defined to use multi-core processing. See
\code{\link[parallel]{clusterSetRNGStream}} for further information.
}
