% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{collect}
\alias{collect}
\title{Collect various population comparison statistics}
\usage{
collect(results, parameters, condition)
}
\arguments{
\item{results}{a list containing the simulation results from compute(), where each cell
is stored in a unique list element. If compute() was modified to return a list instead, then
results will be a list containing list elements. Otherwise, it will be a list containing
numeric elements (the default)}

\item{parameters}{a list containing all the 'parameters' elements generated from sim(), where
each repetition is stored in a unique list element}

\item{condition}{a single row from the Design input (as a data.frame), indicating the
  simulation conditions}
}
\value{
must return a named numeric vector with the desired meta-simulation results
}
\description{
This collapses across the simulation results within each condition for computing composite
estimates such as RMSD, bias, Type I error, etc.
}
\examples{
\dontrun{

mycollect <- function(results, parameters, condition){

    # handy functions
    bias <- function(observed, population) mean(observed - population)
    RMSD <- function(observed, population) sqrt(mean((observed - population)^2))

    #convert to matrix for convenience (if helpful)
    cell_results <- do.call(rbind, results)

    # silly test for bias and RMSD of a random number from 0
    pop_value <- 0
    bias.random_number <- bias(sapply(parameters, function(x) x$random_number), pop_value)
    RMSD.random_number <- RMSD(sapply(parameters, function(x) x$random_number), pop_value)

    #find results of interest here (alpha < .1, .05, .01)
    nms <- c('welch', 'independent')
    lessthan.10 <- colMeans(cell_results[,nms] < .10)
    lessthan.05 <- colMeans(cell_results[,nms] < .05)
    lessthan.01 <- colMeans(cell_results[,nms] < .01)

    # return the results that will be appended to the Design input
    ret <- c(bias.random_number=bias.random_number,
             RMSD.random_number=RMSD.random_number,
             lessthan.10=lessthan.10,
             lessthan.05=lessthan.05,
             lessthan.01=lessthan.01)
    return(ret)
}

}
}

