% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{compute}
\alias{compute}
\title{Compute estimates and statistics}
\usage{
compute(simlist, condition, edit)
}
\arguments{
\item{simlist}{a list containing the data.frame ('dat') and parameters ('parameters')
generated from the sim() function}

\item{condition}{a single row from the Design input (as a data.frame), indicating the
simulation conditions}

\item{edit}{a character indicating whether the function should be debugged}
}
\value{
returns a named numeric vector with the values of interest (e.g., p-values,
  effects sizes, etc), or a list containing values of interest (e.g., seperate matrix
  and vector of parameter estimates corresponding to elements in simlist$parameters)
}
\description{
Compute all relevant test statistics and parameter estimates here.
This is the computational heavy lifting section. In some cases, it may be easier to change
the output to a named list containing different parameter configurations (i.e., when
determining RMSD values for a large set of population parameters).
}
\examples{
\dontrun{

mycompute <- function(simlist, condition, edit){

    if(edit == 'compute') browser()

    # require packages/define functions if needed, or better yet index with the :: operator
    require(stats)
    mygreatfunction <- function(x) print('Do some stuff')

    # begin extracting the data elements
    dat <- simlist$dat
    parameters <- simlist$parameters

    #wrap computational statistics in try() statements to control estimation problems
    welch <- try(t.test(DV ~ group, dat), silent=TRUE)
    ind <- try(stats::t.test(DV ~ group, dat, var.equal=TRUE), silent=TRUE)

    # check if error, and if so stop and return an 'error'. This will re-draw the data
    if(is(welch, 'try-error')) stop('Welch error message')
    if(is(ind, 'try-error')) stop('Independent t-test error message')

    # In this function the p values for the t-tests are returned,
    #  and make sure to name each element, for future reference
    ret <- c(welch = welch$p.value,
             independent = ind$p.value)

    return(ret)
}

}
}

