% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gen_seeds}
\alias{gen_seeds}
\title{Generate random seeds}
\usage{
gen_seeds(design, nsets = 1L)
}
\arguments{
\item{design}{design matrix that requires a unique seed per condition}

\item{nsets}{if separating the simulation across independent nodes manually
then the number of independent sets can be specified to generate a matrix
of seeds. This will return a matrix object with \code{nsets} columns to
be indexed column-wise for each manual seed specification}
}
\description{
Generate seeds to be passed to \code{runSimulation}'s \code{seed} input. Values
are sampled from 1 to 2147483647.
}
\examples{

design <- createDesign(factorA=c(1,2,3),
                       factorB=letters[1:3])
seeds <- gen_seeds(design)
seeds

# two distinct sets of seeds
multi_seeds <- gen_seeds(design, nsets=2)
multi_seeds  # index column-wise for runSimulation(..., seed)

}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
