% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{main}
\alias{main}
\title{Main subroutine}
\usage{
main(index, condition, sim, compute)
}
\arguments{
\item{index}{an integer place-holder value indexed from the 1:each input}

\item{condition}{a single row from the Design input (as a data.frame), indicating the
simulation conditions}

\item{sim}{the sim() function defined above (required for parallel computing)}

\item{compute}{the compute() function defined above (required for parallel computing)}
}
\value{
must return a named list with the 'result' and 'parameters' elements for the
  computational results and Monte Carlo parameters from sim()
}
\description{
This function runs one condition at a time exactly once. This is for
repeating the simulation for each condition a number of times, and is where the
Monte Carlo flow is controlled. Generally speaking, you shouldn't need to edit this
function.
}
\examples{
\dontrun{

# This is the default main function
print(SimDesign::main)

}
}

