% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{MSE}
\alias{MSE}
\title{Compute the mean square error}
\usage{
MSE(observed, population)
}
\arguments{
\item{observed}{a numeric vector of parameter estimates, where the length is equal to the number of
  replications}

\item{population}{a numeric scalar indicating the fixed population value}
}
\value{
returns a single number indicating the overall bias in the estimates
}
\description{
Computes the average deviation (mean square error; also known as the root mean square deviation)
of a sample estimate from the population value.
}
\examples{
\dontrun{

pop <- 1
samp <- rnorm(100, 1, sd = 0.5)
MSE(samp, pop)

}
}

