% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nc}
\alias{nc}
\title{Auto-named Concatenation of Vector or List}
\usage{
nc(..., use.names = FALSE, error.on.duplicate = TRUE)
}
\arguments{
\item{...}{objects to be concatenated}

\item{use.names}{logical indicating if \code{names} should be preserved (unlike \code{\link{c}},
default is \code{FALSE})}

\item{error.on.duplicate}{logical; if the same object name appears in the returning object
should an error be thrown? Default is \code{TRUE}}
}
\description{
This is a wrapper to the function \code{\link{c}}, however names the respective elements
according to their input object name.
}
\examples{

A <- 1
B <- 2
C <- 3

names(C) <- 'LetterC'

# compare the following
c(A, B, C) # unnamed

nc(A, B, C) # named
nc(this=A, B, C) # respects override named (same as c() )
nc(this=A, B, C, use.names = TRUE) # preserve original name

# throws errors if names not unique
nc(this=A, this=B, C)
nc(LetterC=A, B, C, use.names=TRUE)

# poor input choice names
nc(t.test(c(1:2))$p.value, t.test(c(3:4))$p.value)

# better to explicitly provide name
nc(T1 = t.test(c(1:2))$p.value,
   T2 = t.test(c(3:4))$p.value)

# List input
A <- list(1)
B <- list(2:3)
C <- list('C')

names(C) <- 'LetterC'

# compare the following
c(A, B, C) # unnamed

nc(A, B, C) # named
nc(this=A, B, C) # respects override named (same as c() and list() )
nc(this=A, B, C, use.names = TRUE) # preserve original name


}
\references{
Chalmers, R. P., & Adkins, M. C.  (2020). Writing Effective and Reliable Monte Carlo Simulations
with the SimDesign Package. \code{The Quantitative Methods for Psychology, 16}(4), 248-280.
\doi{10.20982/tqmp.16.4.p248}

Sigal, M. J., & Chalmers, R. P. (2016). Play it again: Teaching statistics with Monte
Carlo simulation. \code{Journal of Statistics Education, 24}(3), 136-156.
\doi{10.1080/10691898.2016.1246953}
}
