% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{notify}
\alias{notify}
\title{Send a simulation notification}
\usage{
notify(notifier, event, event_data)
}
\arguments{
\item{notifier}{The notifier object}

\item{event}{Character string indicating the notification trigger ("condition" or "complete")}

\item{event_data}{List containing context information for the notification}
}
\description{
Package extensions can implement custom notifiers by creating S3 methods
for this generic.
}
\keyword{internal}
