% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgenerate.R
\name{rdunif}
\alias{rdunif}
\title{Generate data with the discrete uniform distribution}
\usage{
rdunif(n, min = 1, max = 100)
}
\arguments{
\item{n}{number of observations to generate}

\item{min}{lowest possible integer number to be sampled. Default is 1}

\item{max}{highest possible integer number to be sampled. Default is 100}
}
\value{
a numeric vector
}
\description{
Function generates data from the discrete uniform distribution.
Default draws integer numbers uniformly within the range [1,100].
}
\examples{

# symmetric values between [1,100]
x <- rdunif(1000)
hist(x, 15)

# values between [50,80]
x <- rdunif(1000, min=50, max=80)
hist(x, 15)

}
\references{
Sigal, M. J., & Chalmers, R. P. (2016). Play it again: Teaching statistics with Monte
Carlo simulation. \code{Journal of Statistics Education, 24}(3), 136-156.
\url{http://www.tandfonline.com/doi/full/10.1080/10691898.2016.1246953}
}
\seealso{
\code{\link{runSimulation}}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
