% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{RE}
\alias{RE}
\title{Compute the relative efficiency of multiple estimators}
\usage{
RE(MSEs)
}
\arguments{
\item{MSEs}{a vector of mean square error values (see \code{\link{MSE}})}
}
\value{
returns a vector of ratios indicating the relative efficiency compared to the first
  estimator (which by default will be equal to 1). Values less than 1 indicate worse efficiency, while
  values greater than 1 indicate better efficiency
}
\description{
Computes the relateive efficiency given the MSE values for multiple estimators
}
\examples{
\dontrun{

pop <- 1
samp1 <- rnorm(100, 1, sd = 0.5)
MSE1 <- MSE(samp1, pop)
samp2 <- rnorm(100, 1, sd = 1)
MSE2 <- MSE(samp2, pop)

RE(c(MSE1, MSE2))

}
}

