% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descript.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{select}
\alias{group_by}
\alias{arrange}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:arrange]{arrange()}}, \code{\link[dplyr:group_by]{group_by()}}, \code{\link[dplyr:select]{select()}}}
}}

