% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{rgh}
\alias{rgh}
\title{Generate data with the g-and-h distribution}
\usage{
rgh(n, g, h, a = 0, b = 1)
}
\arguments{
\item{n}{sample size}

\item{g}{the g parameter which controls the skew of a distribution in terms of both direction
and magnitude}

\item{h}{the h parameter which controls the tail weight or elongation of a distribution and
is positively related with kurtosis}

\item{a}{location parameter}

\item{b}{scale parameter (must be positive)}
}
\description{
The g-and-h distriubtion is of the form \code{G(p,g,h,a,b) = a + b * exp((g*Zp)-1)/g * exp((h*Zp^2)/2)},
and can be used to generate several different types of univariate distribution functions.
}
\examples{

set.seed(1)
norm <- rgh(10000,1e-5,0)
hist(norm)

skew <- rgh(10000,1/2,0)
hist(skew)

neg_skew_platykurtic <- rgh(10000,-1,-1/2)
hist(neg_skew_platykurtic)

}
\author{
Phil Chalmers
}
