% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgenerate.R
\name{rtriangle}
\alias{rtriangle}
\title{Generate data with the triangle distribution}
\usage{
rtriangle(n, min = 0, max = 1, middle = (min + max)/2)
}
\arguments{
\item{n}{number of observations to generate}

\item{min}{lower bound location. Default is 0}

\item{max}{upper bound location. Default is 1}

\item{middle}{middle location. Default is (min + max)/2}
}
\value{
a numeric vector
}
\description{
Function generates data from the triangle distribution. Default draws data from a symmetric
triangle distribution within the range [0,1].
}
\examples{

# symmetric values between [0,1]
x <- rtriangle(1000)
hist(x, 15)

# values between [0,2], completely negatively skewed
x <- rtriangle(1000, max=2, middle=2)
hist(x, 15)

}
\references{
Sigal, M. J., & Chalmers, R. P. (2016). Play it again: Teaching statistics with Monte
Carlo simulation. \code{Journal of Statistics Education, 24}(3), 136-156.
\url{http://www.tandfonline.com/doi/full/10.1080/10691898.2016.1246953}
}
\seealso{
\code{\link{runSimulation}}
}
