% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimDesign.R
\docType{package}
\name{SimDesign}
\alias{SimDesign}
\alias{SimDesign-package}
\title{Simulation Design}
\description{
Simulation Design
}
\details{
Functions to organize simulations in R. Tries to manage the structure and back-end of the simulations,
managing issues like resimulating non-convergent results, organizing the functional workflow,
handling parallel back-end information, saving temporary files, and so on. The primary function for
organizing the simulations is \code{\link{runSimulation}}.
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

