% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{SimRead}
\alias{SimRead}
\title{Read simulation files}
\usage{
SimRead(filename)
}
\arguments{
\item{filename}{name of the file to read in}
}
\value{
the R binary object
}
\description{
Convenience function that switches between \code{readRDS()} and
\code{qs2::qs_read()} for functions saved with \code{SimDesign}.
By convention, objects saved with the extension \code{.rds} are read
as R binary files and typically reflect the final object from
\code{\link{runSimulation}} or \code{\link{runArraySimulation}},
while files without an extension are read using \code{qs2} (most often
temporary files or results written to an associated sub-directory).
}
