% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimResults.R
\name{SimResults}
\alias{SimResults}
\title{Function to read in saved simulation results}
\usage{
SimResults(results, which, wd = getwd())
}
\arguments{
\item{results}{object returned from \code{\link{runSimulation}} where \code{save_results = TRUE}
was used}

\item{which}{a numeric vector indicating which rows should be read in. If missing, all rows will be
read in}

\item{wd}{working directory; default is found with \code{\link{getwd}}.}
}
\description{
If \code{\link{runSimulation}} was passed the flag \code{save_results = TRUE} then the
row results corresponding to the \code{design} object will be stored to a suitable
sub-directory as individual \code{.rds} files. While users could use \code{\link{readRDS}} directly
to read these files in themselves, this conviencience function will read the desired rows in
automatically given the returned object
from the simulation. Can be used to read in 1 or more .rds files at once (if more than 1 file
is read in then the result will be stored in a list).
}
\examples{

\dontrun{

results <- runSimulation(..., save_results = TRUE)

# row 1 results
row1 <- SimResults(results, 1)

# rows 1:5, stored in a named list
rows_1to5 <- SimResults(results, 1:5)

# all results
rows_all <- SimResults(results)

}
}

