% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregate_simulations.R
\name{aggregate_simulations}
\alias{aggregate_simulations}
\title{Collapse seperate simulation files into a single result}
\usage{
aggregate_simulations()
}
\value{
a data.frame with the (weighted) average of the simulation results
}
\description{
This function grabs all .rds files in the working directory and aggregates them into a single
data.frame object. Weights are infered from the last numbers before .rds, in the form
"name_of_sim_100.rds". Note that the working direction should ONLY consist of the .rds files to
be aggregated; all other .rds files should be removed or relocated.
}
\examples{
\dontrun{

setwd('my_working_directory')

final <- aggregate_simulations()
head(final)

saveRDS(final, 'my_final_simulation.rds')

}
}

