% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_simulations.R
\name{aggregate_simulations}
\alias{aggregate_simulations}
\title{Collapse separate simulation files into a single result}
\usage{
aggregate_simulations(
  files = NULL,
  filename = NULL,
  dirs = NULL,
  results_dirname = "SimDesign_aggregate_results",
  select = NULL
)
}
\arguments{
\item{files}{a \code{character} vector containing the names of the simulation's final \code{.rds} files}

\item{filename}{(optional) name of .rds file to save aggregate simulation file to. If not specified
then the results will only be returned in the R console}

\item{dirs}{a \code{character} vector containing the names of the \code{save_results} directories to be
aggregated. A new folder will be created and placed in the \code{results_dirname} output folder}

\item{results_dirname}{the new directory to place the aggregated results files}

\item{select}{a character vector indicating columns to variables to select from the
\code{SimExtract(what='results')} and summary information.
This is mainly useful when RAM is an issue
given simulations with many stored estimates. Default includes the results objects
in their entirety, though to omit all internally stored simulation results pass the
character \code{'NONE'}}
}
\value{
if \code{files} is used the function returns a \code{data.frame/tibble} with the (weighted) average
  of the simulation results. Otherwise, if \code{dirs} is used, the function returns NULL
}
\description{
This function aggregates the results from SimDesign's \code{\link{runSimulation}} into a single
objects suitable for post-analyses, or combines all the saved results directories and combines
them into one. This is useful when results are run piecewise on one node (e.g., 500 replications
in one batch, 500 again at a later date) or run independently across different
nodes/computers that are not on the same network.
}
\examples{
\dontrun{

setwd('my_working_directory')

## run simulations to save the .rds files (or move them to the working directory)
# ret1 <- runSimulation(..., filename='file1')
# ret2 <- runSimulation(..., filename='file2')

# saves to the hard-drive and stores in workspace
final <- aggregate_simulations(files = c('file1.rds', 'file2.rds'))
final

# If filename not included, can be extracted from results
# files <- c(SimExtract(ret1, 'filename'), SimExtract(ret2, 'filename'))
# final <- aggregate_simulations(files = files)

# aggregate saved results for .rds files and results directories
# runSimulation(..., save_results = TRUE, save_details = list(save_results_dirname = 'dir1'))
# runSimulation(..., save_results = TRUE, save_details = list(save_results_dirname = 'dir2'))

# place new saved results in 'SimDesign_results/' by default
aggregate_simulations(files = c('file1.rds', 'file2.rds'),
                      filename='aggreged_sim.rds',
                      dirs = c('dir1', 'dir2'))

# If dirnames not included, can be extracted from results
# dirs <- c(SimExtract(ret1, 'save_results_dirname'),
            SimExtract(ret2, 'save_results_dirname'))
# aggregate_simulations(dirs = dirs)

#################################################
# Example where simulation run in multiple batches

Design <- createDesign(N  = c(30, 60),
                       mu = c(0,5))

# want 1000 replications in total, but execute two batches
replications <- 1000 / 2

#-------------------------------------------------------------------

Generate <- function(condition, fixed_objects = NULL) {
    dat <- with(condition, rnorm(N, mean=mu))
    dat
}

Analyse <- function(condition, dat, fixed_objects = NULL) {
    ret <- c(mean=mean(dat), SD=sd(dat))
    ret
}

Summarise <- function(condition, results, fixed_objects = NULL) {
    ret <- colMeans(results)
    ret
}

#-------------------------------------------------------------------

# Generate fixed seeds to be distributed (two sets of seeds)
set.seed(4321)
seeds <- gen_seeds(Design, nsets=2)
seeds

# distribute both jobs independently (total reps / 2 )
runSimulation(design=Design, replications = 1000 / 2,
                 generate=Generate, analyse=Analyse, summarise=Summarise,
                 filename='job-1', seed=seeds[,1])

runSimulation(design=Design, replications = 1000 / 2,
                 generate=Generate, analyse=Analyse, summarise=Summarise,
                 filename='job-2', seed=seeds[,2])

# aggregate into single object, and save to drive as 'collected.rds'
sim <- aggregate_simulations(files = c('job-1.rds', 'job-2.rds'),
                             filename='collected.rds')
sim
file.exists('collected.rds')


#################################################
# Example where each row condition is repeated, evaluated independently,
# and later collapsed into a single analysis object

# Each condition repeated four times (hence, replications
# should be set to desired.reps/4)
Design <- createDesign(N  = c(30, 60),
                       mu = c(0,5),
                       repeat_conditions = 4L)

#-------------------------------------------------------------------

# Generate-Analyse-Summarise same as previous example

#-------------------------------------------------------------------

# Generate fixed seeds to be distributed
set.seed(1234)
seeds <- gen_seeds(Design)
seeds

# replications vector (constant is fine if the same across conditions;
# below is vectorized to demonstrate that this could change)
replications <- rep(250, nrow(Design))

# create directory to store all final simulation files
dir.create('sim_files/')

# distribute jobs independently (explicitly parallelize here on cluster)
sapply(1:nrow(Design), \(i) {
  runSimulation(design=Design[i, ], replications=replications[i],
                generate=Generate, analyse=Analyse, summarise=Summarise,
                filename=paste0('sim_files/job-', i))
})

# aggregate into single object
sim <- aggregate_simulations(files = paste0('sim_files/job-',
                                     1:nrow(Design), ".rds"))
sim

}
}
\references{
Chalmers, R. P., & Adkins, M. C.  (2020). Writing Effective and Reliable Monte Carlo Simulations
with the SimDesign Package. \code{The Quantitative Methods for Psychology, 16}(4), 248-280.
\doi{10.20982/tqmp.16.4.p248}

Sigal, M. J., & Chalmers, R. P. (2016). Play it again: Teaching statistics with Monte
Carlo simulation. \code{Journal of Statistics Education, 24}(3), 136-156.
\doi{10.1080/10691898.2016.1246953}
}
\seealso{
\code{\link{runSimulation}}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
