% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyse}
\alias{analyse}
\title{Compute estimates and statistics}
\usage{
analyse(condition, dat, fixed_objects = NULL, parameters = NULL)
}
\arguments{
\item{condition}{a single row from the design input (as a \code{data.frame}), indicating the
simulation conditions}

\item{dat}{the \code{dat} object returned from the \code{\link{generate}} function
(usually a \code{data.frame}, \code{matrix}, or \code{vector}).}

\item{fixed_objects}{object passed down from \code{\link{runSimulation}}}

\item{parameters}{the (optional) list object named 'parameters' returned from the
\code{\link{generate}} function when a list is returned. Otherwise, this will be an just an
empty list}
}
\value{
returns a named \code{numeric} vector with the values of interest (e.g., p-values,
  effects sizes, etc), or a \code{list} containing values of interest (e.g., separate matrix
  and vector of parameter estimates corresponding to elements in \code{parameters})
}
\description{
Computes all relevant test statistics, parameter estimates, detection rates, and so on.
This is the computational heavy lifting portion of the Monte Carlo simulation.
}
\details{
In some cases, it may be easier to change the output to a named \code{list} containing
different parameter configurations (e.g., when
determining RMSE values for a large set of population parameters).

The use of \code{\link{try}} functions is generally not required because the function
is internally wrapped in a \code{\link{try}} call. Therefore, if a function stops early
then this will cause the function to halt iternally, the message which triggered the \code{\link{stop}}
will be recorded, and \code{\link{generate}} will be called again to obtain a different dataset.
That being said, it may be useful for users to throw their own \code{\link{stop}} commands if the data
should be redrawn for other reasons (e.g., a model terminated correctly but the maximum number of
iterations were reached).
}
\examples{
\dontrun{

myanalyse <- function(condition, dat, fixed_objects = NULL, parameters = NULL){

    # require packages/define functions if needed, or better yet index with the :: operator
    require(stats)
    mygreatfunction <- function(x) print('Do some stuff')

    #wrap computational statistics in try() statements to control estimation problems
    welch <- t.test(DV ~ group, dat)
    ind <- stats::t.test(DV ~ group, dat, var.equal=TRUE)

    # In this function the p values for the t-tests are returned,
    #  and make sure to name each element, for future reference
    ret <- c(welch = welch$p.value,
             independent = ind$p.value)

    return(ret)
}

}
}
\seealso{
\code{\link{stop}}
}

