% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDesign.R
\docType{data}
\name{BF_sim_alternative}
\alias{BF_sim_alternative}
\title{(Alternative) Example simulation from Brown and Forsythe (1974)}
\description{
Example results from the Brown and Forsythe (1974) article on robust estimators for
variance ratio tests. Statistical tests and distributions are organized by columns
and the unique design conditions are organized by rows. See \code{\link{BF_sim}} for an alternative
form of the same simulation where distributions are also included in the rows.
Code for this simulation is available on the wiki (\url{https://github.com/philchalmers/SimDesign/wiki}).
}
\examples{

\dontrun{
data(BF_sim_alternative)
head(BF_sim_alternative)

#' #Type I errors
subset(BF_sim_alternative, var_ratio == 1)
}
}
\references{
Brown, M. B. and Forsythe, A. B. (1974). Robust tests for the equality of variances.
  \emph{Journal of the American Statistical Association, 69}(346), 364--367.

Chalmers, R. P., & Adkins, M. C.  (2020). Writing Effective and Reliable Monte Carlo Simulations
with the SimDesign Package. \code{The Quantitative Methods for Psychology, 16}(4), 249-281.
\doi{10.20982/tqmp.16.4.p249}

Sigal, M. J., & Chalmers, R. P. (2016). Play it again: Teaching statistics with Monte
Carlo simulation. \code{Journal of Statistics Education, 24}(3), 136-156.
\doi{10.1080/10691898.2016.1246953}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}
