% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{bias}
\alias{bias}
\title{Compute bias summary statistic}
\usage{
bias(observed, population = NULL)
}
\arguments{
\item{observed}{a numeric vector of parameter estimates, where the length is equal to the number of
  replications}

\item{population}{a numeric scalar indicating the fixed population value. If NULL, then it will be assumed
  that the \code{observed} input is in a deviation form (therefore \code{mean(observed)} will be returned)}
}
\value{
returns a single number indicating the overall bias in the estimates
}
\description{
Computes the bias of a sample estimate from the population value. Accepts observed and population values,
as well as observed values which are in deviation form.
}
\examples{
\dontrun{

pop <- 1
samp <- rnorm(100, 1, sd = 0.5)
bias(samp, pop)

dev <- samp - pop
bias(dev)

}
}

