% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bradley1978.R
\name{Bradley1978}
\alias{Bradley1978}
\title{Bradley's (1978) empirical robustness interval}
\usage{
Bradley1978(
  p,
  alpha = 0.05,
  type = "liberal",
  unname = FALSE,
  out.logical = FALSE,
  out.labels = c("conservative", "robust", "liberal")
)
}
\arguments{
\item{p}{(optional) numeric vector containing the empirical detection
rate(s) estimates. If supplied a character vector with elements defined in
\code{out.labels} or a logical vector will be returned indicating whether the
detection rate estimate is considered 'robust'. If missing, the
interval criteria will be printed to the console}

\item{alpha}{Type I error rate to evaluated (default is .05)}

\item{type}{character vector indicating the type of interval classification to use.
Default is 'liberal', however can be 'stringent' to use Bradley's more
stringent robustness criteria}

\item{unname}{logical; apply \code{\link{unname}} to the results to remove any variable
names?}

\item{out.logical}{logical; should the output vector be TRUE/FALSE indicating whether
the supplied empirical detection rate should be considered "robust"? Default is
FALSE, in which case the out.labels elements are used instead}

\item{out.labels}{character vector of length three indicating the classification
labels according to the desired robustness interval}
}
\description{
Robustness interval criteria for empirical detection rate estimates defined
by Bradley (1978). See \code{\link{EDR}} to obtain such estimates.
}
\examples{

# interval criteria used
Bradley1978()
Bradley1978(type = 'stringent')
Bradley1978(alpha = .01, type = 'stringent')

# intervals applied to empirical detection rate estimates
edr <- c(test1=.05, test2=.027, test3=.051, test4=.076, test5=.024)

Bradley1978(edr)
Bradley1978(edr, out.logical=TRUE) # is robust?

}
\references{
Bradley, J. V. (1978). Robustness? \emph{British Journal of Mathematical and
Statistical Psychology, 31}, 144-152.

Chalmers, R. P., & Adkins, M. C.  (2020). Writing Effective and Reliable Monte Carlo Simulations
with the SimDesign Package. \code{The Quantitative Methods for Psychology, 16}(4), 248-280.
\doi{10.20982/tqmp.16.4.p248}

Sigal, M. J., & Chalmers, R. P. (2016). Play it again: Teaching statistics with Monte
Carlo simulation. \code{Journal of Statistics Education, 24}(3), 136-156.
\doi{10.1080/10691898.2016.1246953}
}
\seealso{
\code{\link{EDR}}, \code{\link{Serlin2000}}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
