% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{ECR}
\alias{ECR}
\title{Compute the empirical coverage rate for Type I errors and Power}
\usage{
ECR(CIs, parameter, tails = FALSE)
}
\arguments{
\item{CIs}{a \code{numeric} vector or \code{matrix} of confidence interval values for a
given parameter value, where the first element/column indicates the lower confidence interval
and the second element/column the upper confidence interval. If a
vector of length 2 is passed instead then the returned value will be either a 1 or 0 to indicate
whether the parameter value was or was not within the interval, respectively}

\item{parameter}{a numeric scalar indicating the fixed parameter value. Alternative, a \code{numeric}
vector object with length equal to the number of rows as \code{CIs} (use to compare sets of parameters
at once)}

\item{tails}{logical; when TRUE returns a vector of length 2 to indicate the proportion of times
the parameter was lower or higher than the supplied interval, respectively. This is mainly only
useful when the coverage region is not expected to be symmetric, and therefore is generally not
required. Note that \code{1 - sum(ECR(CIs, parameter, tails=TRUE)) == ECR(CIs, parameter)}}
}
\description{
Computes the detection rate for determining empirical Type I error and power rates using information
from the confidence intervals. Note that using \code{1 - ECR(CIs, parameter)} will provide the empirical
detection rate.
}
\examples{

CIs <- matrix(NA, 100, 2)
for(i in 1:100){
   dat <- rnorm(100)
   CIs[i,] <- t.test(dat)$conf.int
}

ECR(CIs, 0)
ECR(CIs, 0, tails = TRUE)

# single vector input
CI <- c(-1, 1)
ECR(CI, 0)
ECR(CI, 2)
ECR(CI, 2, tails = TRUE)

# parameters of the same size as CI
parameters <- 1:10
CIs <- cbind(parameters - runif(10), parameters + runif(10))
parameters <- parameters + rnorm(10)
ECR(CIs, parameters)

}
\seealso{
\code{\link{EDR}}
}

