% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{EDR}
\alias{EDR}
\title{Compute the empirical detection rate for Type I errors and Power}
\usage{
EDR(p, alpha)
}
\arguments{
\item{p}{a vector or matrix of p-values from the desired statistical estimator. If a matrix,
  each statistic must be organized by column where the number of rows is equal to the number
  of replications}

\item{alpha}{the nominal detection rate to be studied (typical values are .10, .05, and .01)}
}
\description{
Computes the detection rate for determining empirical Type I error and power rates
}
\examples{
\dontrun{

rates <- numeric(100)
for(i in 1:100){
   dat <- rnorm(100)
   rates[i] <- t.test(dat)$p.value
}

EDR(rates, alpha = .05)

}
}

