% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_section.R
\name{gold_section}
\alias{gold_section}
\title{Gold section search algorithm}
\usage{
gold_section(f, interval, integer = FALSE, tol = 0.001, maxiter = 100L, ...)
}
\arguments{
\item{f}{function to be minimized, where the first argument corresponds
to the parameter of interest}

\item{interval}{interval with which to search, specified as
\code{c(lower, upper)}}

\item{integer}{logical; should the parameters be constrained as
integer values?}

\item{tol}{convergence tolerance. Ignored when \code{integer = TRUE}
(converges when lowest integer is found between competing boundaries)}

\item{maxiter}{maximum number of iterations}

\item{...}{additional arguments passed to \code{f}}
}
\value{
a \code{list} of the converge and solution information
}
\description{
Implementation of univariate gold-section search algorithm. Algorithm
assumes that the objective function has one unique minimum (unimodal)
within the specified search interval. Supports integer searches for input
parameter.
}
\examples{

y <- c(10, 20, 50)
y <- y - mean(y)
x <- 1:3
f <- function(beta, y, x) sum( (y - beta*x)^2)

betas <- seq(0, 10, length.out=1000)
fs <- sapply(betas, f, y=y, x=x)
plot(betas, fs, type = 'l', las=1)

res <- gold_section(f, interval = c(0,10), y=y, x=x)
res

# if beta must be an integer
betas <- -10:20
fs <- sapply(betas, f, y=y, x=x)
plot(betas, fs, las=1)

res <- gold_section(f, interval = c(-10,20), integer=TRUE, y=y, x=x)
res

}
\references{
Chalmers, R. P., & Adkins, M. C.  (2020). Writing Effective and Reliable Monte Carlo Simulations
with the SimDesign Package. \code{The Quantitative Methods for Psychology, 16}(4), 248-280.
\doi{10.20982/tqmp.16.4.p248}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
