% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{MAE}
\alias{MAE}
\title{Compute the mean absolute error}
\usage{
MAE(estimate, parameter = NULL, type = "MAE")
}
\arguments{
\item{estimate}{a numeric vector or matrix/data.frame of parameter estimates. If a vector,
the length is equal to the number of replications. If a matrix/data.frame,
the number of rows must equal the number of replications}

\item{parameter}{a numeric scalar/vector indicating the fixed parameter values.
If a single value is supplied and \code{estimate} is a matrix/data.frame then the value will be
recycled for each column.
If NULL, then it will be assumed that the \code{estimate} input is in a deviation
form (therefore \code{mean(abs(estimate))} will be returned)}

\item{type}{type of deviation to compute. Can be 'MAE' (default) for the mean absolute error, or
'NMSE' for the normalized MAE (MAE / (max(estimate) - min(estimate)))}
}
\value{
returns a numeric vector indicating the overall mean absolute error in the estimates
}
\description{
Computes the average absolute deviation of a sample estimate from the parameter value.
Accepts estimate and parameter values,
as well as estimate values which are in deviation form.
}
\examples{

pop <- 1
samp <- rnorm(100, 1, sd = 0.5)
MAE(samp, pop)

dev <- samp - pop
MAE(dev)
MAE(samp, pop, type = 'NMAE')

# matrix input
mat <- cbind(M1=rnorm(100, 2, sd = 0.5), M2 = rnorm(100, 2, sd = 1))
MAE(mat, parameter = 2)

# same, but with data.frame
df <- data.frame(M1=rnorm(100, 2, sd = 0.5), M2 = rnorm(100, 2, sd = 1))
MAE(df, parameter = c(2,2))

}
\seealso{
RMSE
}

