% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{main}
\alias{main}
\title{Main subroutine}
\usage{
main(index, condition, generate, analyse, fixed_design_elements)
}
\arguments{
\item{index}{an integer place-holder value indexed from the 1:each input}

\item{condition}{a single row from the design input (as a data.frame), indicating the
simulation conditions}

\item{generate}{the \code{\link{generate}} function defined above (required for parallel computing)}

\item{analyse}{the \code{\link{analyse}} function defined above (required for parallel computing)}

\item{fixed_design_elements}{object passed down from \code{\link{runSimulation}}}
}
\value{
must return a named list with the 'result' and 'parameters' elements for the
  computational results and \code{parameters} from \code{\link{generate}}
}
\description{
This function runs one condition at a time exactly once. This is for
repeating the simulation for each condition a number of times, and is where the
Monte Carlo flow is controlled. Generally speaking, you shouldn't need to edit this
function.
}
\examples{
\dontrun{

# This is the default main function
print(SimDesign::main)

}
}

