% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Run a Monte Carlo simulation given a data.frame of conditions and simulation functions}
\usage{
runSimulation(design, replications, generate, analyse, summarise,
  parallel = FALSE, MPI = FALSE, save = FALSE, save_every = 1,
  clean = TRUE, compname = Sys.info()["nodename"],
  filename = paste0(compname, "_Final_", replications, ".rds"),
  tmpfilename = paste0(compname, "_tmpsim.rds"), main = NULL,
  ncores = parallel::detectCores(), edit = "none", verbose = TRUE)
}
\arguments{
\item{design}{a data.frame object containing the Monte carlo simulation conditions to
  be studied, where each row represents a unique condition}

\item{replications}{number of replication to perform per condition (i.e., each row in \code{design})}

\item{generate}{user-defined data and parameter generating function.
  See \code{\link{generate}} for details}

\item{analyse}{user-defined computation function which acts on the dat generated from
  \code{\link{generate}}. See \code{\link{analyse}} for details}

\item{summarise}{user-defined summary function to be used after all the replications have completed.
   See \code{\link{summarise}} for details}

\item{parallel}{logical; use parallel processing from the \code{parallel} package over each
  unique condition?

  NOTE: if using other packages other than the base attached ones (e.g., \code{stats},
  \code{graphics}, \code{utils}, etc) then you must either a) explicitly load the packages within
  the respective defined functions with a \code{library()} or \code{require()} call, or b) use the
  :: operator to locate the functions that are not attached by default}

\item{MPI}{logical; use the \code{doMPI} package to run simulation in parallel on
  a cluster? Default is FALSE}

\item{save}{logical; save the final simulation and temp files to the hard-drive? Default is FALSE}

\item{save_every}{a number indicating how often to temporarily save your simulation results to
  disk. Default is 1 to save after every condition is complete, but set to NA if you don't
  want to save any temp files}

\item{clean}{logical; remove any temp files that are created after the simulation is complete?
  Default is TRUE}

\item{compname}{name of the computer running the simulation. Normally this doesn't need to be modified,
  but in the event that a node breaks down while running a simulation the results from the tmp files
  may be resumed on another computer by changing the name of the node to match the broken computer}

\item{filename}{the name of the file to save the results to. Default is the system name with
  the number of replications and 'Final' appended to the string}

\item{tmpfilename}{the name of the temporary file, default is the system name with 'tmpsim.rds'
  appended at the end. This file will be
  read in if it is in the working directory, and the simulation will continue where at the last
  point this file was saved (useful in case of power outages or broken nodes).
  This file will be deleted when the simulation is complete}

\item{main}{(optional) user-defined organization function defining how the simulation should be
   organized. When NULL, the internal function definition is used (and the majority of the time
   this is sufficient). See \code{\link{main}} for further details}

\item{ncores}{number of cores to be used in parallel execution. Default uses all available}

\item{edit}{a string indicating where to initiate a `browser()` call for editing and debugging.
  Options are 'none' (default), 'main' to edit the main function calls loop, 'generate' to edit the
  data simulation function, 'analyse' to edit the computational function, and 'summarise' to
  edit the collection function. Alternatively, users may place \code{\link{browser}} calls within their
  own code for debugging at specific lines (note: parallel computation flags will
  automatically be disabled when this is detected)}

\item{verbose}{logical; print messages to the R console?}
}
\description{
This function runs a Monte Carlo simulation study given the simulation functions, the design conditions,
and the number of replications. Results can be saved as temporary files in case of interruptions
and may be restored by rerunning the exact function calls again, provided that the respective temp
file can be found in the working directory. Supports parallel and cluster computing, global and
local debugging, and is designed to be cross-platform.  For a skeleton version of the workflow
which may be useful when initially defining a simulation, see \code{\link{SimDesign_functions}}.
}
\details{
The strategy for organizing the Monte Carlo simulation work-flow is to

\describe{
   \item{1)}{Define a suitable \code{design} data.frame. This is often expedited by using the
      \code{\link{expand.grid}} function}
   \item{2)}{Define the three step functions to simulate the data (\code{\link{generate}}),
      analyse the generated data by computing the respective parameter estimates, detection rates,
      etc (\code{\link{analyse}}), and finally summarise the results across the total
      number of replications (\code{\link{summarise}})
   }
   \item{3)}{Pass the above objects to the \code{runSimulation} function, and define the
      number of replications with the \code{replications} input}
   \item{4)}{Analyze the output from \code{runSimulation}, possibly using techniques from ANOVA
     and generating suitable plots}
}

Two constants for each condition are returned by default:
N_CELL_RUNS to indicate the number of Monte Carlo runs it took to obtain valid results (this will
be greater than the number of replications requested if, for example, models failed to converge
and had to be re-drawn), and SIM_TIME to indicate how long (in seconds) it took to complete
all the Monte Carlo replications for each respective condition.

In the event of a computer crash, power outage, etc, if the 'save_every' command has been
used to save temporary results then the original code in main.R need only be rerun again.
The temp file will be read into the design, and the simulation will continue where it left
off before the simulation was terminated. Upon completion, a data.frame with the simulation
will be returned in the R session and a '.rds' file will be saved to the hard-drive (with the
file name corresponding to the 'filename' argument below).
}
\section{Cluster computing}{


If the package is installed across a cluster of computers, and all the computers are accessible on
the same LAN network, then the package may be run in cluster mode using the MPI paradigm. This simply
requires that the computers be setup using the usual MPI requirements (typically, running some flavor
of Linux, have password-less openSSH access, addresses have been added to the /etc/hosts file, etc),
and in the code the argument \code{MPI = TRUE} should be passed.

For instances, if the following code is run on the master node through a terminal then 16 processes
will be summoned (1 master, 15 slaves) across the computers named localhost, slave1, and slave2.

mpirun -np 16 -H localhost,slave1,slave2 R --slave -f simulation.R
}

\section{Poor man's cluster computing}{


In the event that you do not have access to a Beowulf-type cluster, but have multiple personal
computers, then the simulation code can be manually distributed across each computer instead.
This simply requires passing a smaller value to the \code{each} argument on each computer, and later
aggregating the results using the \code{\link{aggregate_simulations}} function.

For instance, if you have two computers available and wanted 500 replications you
could pass \code{replications = 300} to one computer and \code{replications = 200} to the other.
This will create two .rds files which can be combined later with the
\code{\link{aggregate_simulations}} function.
}
\examples{
\dontrun{

#### Step 1 --- Define your conditions under study and create design data.frame

# (use EXPLICIT names, avoid things like N <- 100. That's fine in functions, not here)
sample_sizes <- c(10, 20, 50, 100)
standard_deviation_ratios <- c(1, 4, 8)

Design <- expand.grid(sample_size_group1=sample_sizes,
                      sample_size_group2=sample_sizes,
                      standard_deviation_ratio=standard_deviation_ratios)
dim(Design)
head(Design)

#~~~~~~~~~~~~~~~~~~~~~~~~
#### Step 2 --- Define sim, compute, and collect functions

# skeleton functions to be edited
SimDesign_functions()

# help(generate)
Generate <- function(condition){

    #require packages/define functions if needed, or better yet index with the :: operator

    N1 <- condition$sample_size_group1
    N2 <- condition$sample_size_group2
    sd <- condition$standard_deviation_ratio

    group1 <- rnorm(N1)
    group2 <- rnorm(N2, sd=sd)
    dat <- data.frame(group = c(rep('g1', N1), rep('g2', N2)), DV = c(group1, group2))

    return(dat)
}

# help(analyse)

Analyse <- function(dat, parameters, condition){

    # require packages/define functions if needed, or better yet index with the :: operator
    require(stats)
    mygreatfunction <- function(x) print('Do some stuff')

    #wrap computational statistics in try() statements to control estimation problems
    welch <- try(t.test(DV ~ group, dat), silent=TRUE)
    ind <- try(t.test(DV ~ group, dat, var.equal=TRUE), silent=TRUE)

    # check if any errors occured. This will re-draw the data
    check_error(welch, ind)

    # In this function the p values for the t-tests are returned,
    #  and make sure to name each element, for future reference
    ret <- c(welch = welch$p.value, independent = ind$p.value)

    return(ret)
}

# help(summarise)

Summarise <- function(results, parameters_list, condition){

    #find results of interest here (e.g., alpha < .1, .05, .01)
    nms <- c('welch', 'independent')
    lessthan.05 <- EDR(results[,nms], alpha = .05)

    # return the results that will be appended to the design input
    ret <- c(lessthan.05=lessthan.05)
    return(ret)
}


#~~~~~~~~~~~~~~~~~~~~~~~~
#### Step 3 --- Collect results by looping over the rows in design

# this simulation does not save temp files or the final result to disk (save=FALSE)
Final <- runSimulation(design=Design, replications=1000, parallel=TRUE,
                       generate=Generate, analyse=Analyse, summarise=Summarise)

## Debug the generate function (not run). See ?browser for help on debugging
# runSimulation(design=Design, replications=1000,
#               generate=Generate, analyse=Analyse, summarise=Summarise,
#               parallel=TRUE, edit='generate')




## EXTRA: To run the simulation on a MPI cluster, use the following setup on each node (not run)
# library(doMPI)
# cl <- startMPIcluster()
# registerDoMPI(cl)
# Final <- runSimulation(design=Design, replications=1000, MPI=TRUE,
#                        generate=Generate, analyse=Analyse, summarise=Summarise)
# closeCluster(cl)
# mpi.quit()



#~~~~~~~~~~~~~~~~~~~~~~~~
# Step 4 --- Post-analysis: Create a new R file for analyzing the Final data.frame with R based
# regression stuff, so use the lm() function to find main effects, interactions, plots, etc.
# This is where you get to be a data analyst!
head(Final)
# View(Final)

psych::describe(Final)
psych::describeBy(Final, group = Final$standard_deviations)

# define new variables, if they are helpful
Final$SSratio <- with(Final, sample_sizes_group1 / sample_sizes_group2)

# make numerics into factors for lm() analyses
Final$fSSratio <- with(Final, factor(SSratio))
Final$fSD <- with(Final, factor(standard_deviations))

#lm analysis (might want to change DV to a logit for better stability)
mod <- lm(lessthan.05.welch ~ fSSratio * fSD, Final)
car::Anova(mod)

mod2 <- lm(lessthan.05.independent ~ fSSratio * fSD, Final)
car::Anova(mod2)

library(effects)
plot(allEffects(mod2))

# make some plots
library(ggplot2)
library(reshape2)
welch_ind <- Final[,c('SSratio', "standard_deviations",
    "lessthan.05.welch", "lessthan.05.independent")]
dd <- melt(welch_ind, id.vars = names(welch_ind)[1:2])

ggplot(dd, aes(factor(SSratio), value)) +
    geom_abline(intercept=0.05, slope=0, col = 'red') +
    geom_abline(intercept=0.075, slope=0, col = 'red', linetype='dotted') +
    geom_abline(intercept=0.025, slope=0, col = 'red', linetype='dotted') +
    geom_boxplot() + facet_wrap(~variable)
ggplot(dd, aes(factor(SSratio), value, fill = factor(standard_deviations))) +
    geom_abline(intercept=0.05, slope=0, col = 'red') +
    geom_abline(intercept=0.075, slope=0, col = 'red', linetype='dotted') +
    geom_abline(intercept=0.025, slope=0, col = 'red', linetype='dotted') +
    geom_boxplot() + facet_grid(standard_deviations~variable)


}
}
\seealso{
\code{\link{generate}}, \code{\link{analyse}}, \code{\link{summarise}},
  \code{\link{SimDesign_functions}}
}

