% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimAnova.R
\name{SimAnova}
\alias{SimAnova}
\title{Function for decomposing the simulation into ANOVA-based effect sizes}
\usage{
SimAnova(formula, dat, rates = TRUE)
}
\arguments{
\item{formula}{}

\item{dat}{an object returned from \code{\link{runSimulation}}}

\item{rates}{logical; does the dependent variable consist of rates (e.g., returned from
\code{\link{ECR}} or \code{\link{EDR}})? Default is TRUE, which will use the logit of the DV
to help stabalize the proportions when computing the parameters and effect sizes}
}
\value{
returns a single object containing the data to be analyzed (usually a
  \code{vector}, \code{matrix}, or \code{data.frame}),
  or a list with a the elements \code{'dat'} and \code{'parameters'}. If a list is returned
  the \code{'dat'} element should be the observed data object while the
  \code{'parameters'} element should be a named list containing the simulated parameters
  (if there are any. Otherwise, this could just be an empty list)
}
\description{
Given the results from a simulation with \code{\link{runSimulation}} form an ANOVA table (without
p-values) indicating the effect sizes based on the eta-squared statistic. These provide approximate
indications of observable simulation effects, therefore use these results as exploratory rather
than inferential tools.
}
\examples{
\dontrun{

# Given 'Final' object from runSimulation() example, inspect 2 way interactions
Final[1:3] <- lapply(Final[1:3], factor)
SimAnova(lessthan.05.welch ~ (sample_size + group_size_ratio + standard_deviation_ratio)^2, Final)
SimAnova(lessthan.05.welch ~ (sample_size + group_size_ratio + standard_deviation_ratio)^3, Final)

SimAnova(lessthan.05.independent ~ (sample_size + group_size_ratio + standard_deviation_ratio)^2,
  Final)

}

}

