% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimDesign.R
\docType{package}
\name{SimDesign}
\alias{SimDesign}
\alias{SimDesign-package}
\title{Simulation Design}
\description{
Simulation Design
}
\details{
This package provides tools to help organize Monte Carlo simulations in R.
The tools provided control the structure and back-end of the Monte Carlo simulations
by utilizing a generate-analyse-summarise strategy. The package
attempts to handle issues such as re-simulating non-convergent results,
supports parallel back-end computations, saves and restores temporary files,
aggregates results across independent nodes, and provides native support for debugging.
The primary function for
organizing the simulations is \code{\link{runSimulation}}.
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

