% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimDesign.R
\docType{package}
\name{SimDesign}
\alias{SimDesign}
\alias{SimDesign-package}
\title{Simulation Design}
\description{
Simulation Design
}
\details{
Functions to organize Monte Carlo simulations in R. Tries to manage the structure
and back-end Monte Carlo simulations, managing issues like resimulating
non-convergent results, organizing the functional workflow through a
generate -> analyse -> summerise stragegy, handling parallel back-end information,
saving and restoring temporary files, and so on. The primary function for
organizing the simulations is \code{\link{runSimulation}}.
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

