% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimDesign_functions.R
\name{SimDesign_functions}
\alias{SimDesign_functions}
\title{Skeleton functions for simulations}
\usage{
SimDesign_functions(main = FALSE)
}
\arguments{
\item{main}{logical; include a skeleton of the \code{main} function?}
}
\description{
This function prints skeleton versions of the required functions and work-flow required
to run simulations, complete with the correct inputs and class of outputs. Use this at the start
when defining your simulation.
}
\examples{
SimDesign_functions()
}

