% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.SimDesign.R
\name{subset.SimDesign}
\alias{subset.SimDesign}
\title{Subset method for SimDesign objects}
\usage{
\method{subset}{SimDesign}(x, subset, select, drop = FALSE, ...)
}
\arguments{
\item{x}{A \code{data.frame} object, of class \code{SimDesign}.}
}
\value{
A \code{data.frame} object.
}
\description{
\code{subset.SimDesign} is a default method for subsetting a
\code{data.frame} of class \code{SimDesign}. This is a modification
of the base R subset command to maintain the extra attributes
produced during a simulation.
}
\examples{
\dontrun{
data("BF_sim")
x <- subset(BF_sim, select = 1:6)
attributes(x)
}

}
\seealso{
\code{\link{SimDesign}}
}
